refactor(server): move startup timing under util and mark the first project
project/instance.ts importing @/server/startup was the only src/project ->
src/server edge in the tree and inverted the layering. The Startup namespace
only needs util/log, so it lives in util/startup.ts now.

The timing line's `instance` leg measured the server-cwd instance the global
routes mint, not the first project the sidecar evidence singled out. Instance
creation now reports whether the directory is the server cwd, and the line
carries a separate `project` mark for the first instance of any other
directory.
