+++ backend/cli/src/server/server.ts
@@ -68,7 +68,7 @@ import { OnboardingAuthRoutes } from "./routes/onboarding-auth"
 import { AccountRoutes } from "./routes/account"
 import { BillingSettingsRoutes } from "./routes/settings/billing"
 import { WalletSettingsRoutes } from "./routes/settings/wallet"
-import { Startup } from "./startup"
+import { Startup } from "../util/startup"
 
 // @ts-ignore This global is needed to prevent ai-sdk from logging warnings to stdout https://github.com/vercel/ai/blob/2dc67e0ef538307f21368db32d5a12345d98831b/packages/ai/src/logger/log-warnings.ts#L85
 globalThis.AI_SDK_LOG_WARNINGS = false
