+++ backend/cli/test/server/startup-timing.test.ts
@@ -1,22 +1,31 @@
 import { afterEach, expect, test } from "bun:test"
 import { Server } from "../../src/server/server"
-import { Startup } from "../../src/server/startup"
+import { Startup } from "../../src/util/startup"
 
 afterEach(() => Startup.reset())
 
-test("records listening, first instance and interactive once each, in order", () => {
+test("records listening, first instance, first project and interactive once each, in order", () => {
   Startup.reset()
   Startup.listening()
   const listening = Startup.snapshot().listening
   Startup.listening()
   expect(Startup.snapshot().listening).toBe(listening)
 
-  Startup.instance()
+  // The server-cwd instance the global routes mint is not a project.
+  Startup.instance("cwd")
+  const instance = Startup.snapshot().instance
+  expect(instance).toBeGreaterThanOrEqual(listening!)
+  expect(Startup.snapshot().project).toBeUndefined()
+
+  Startup.instance("project")
+  expect(Startup.snapshot().instance).toBe(instance)
+  expect(Startup.snapshot().project).toBeGreaterThanOrEqual(instance!)
+
   Startup.interactive({ page: 5 })
   const marks = Startup.snapshot()
-  expect(marks.instance).toBeGreaterThanOrEqual(listening!)
-  expect(marks.interactive).toBeGreaterThanOrEqual(marks.instance!)
+  expect(marks.interactive).toBeGreaterThanOrEqual(marks.project!)
 
+  Startup.instance("project")
   Startup.interactive({ page: 9 })
   expect(Startup.snapshot()).toEqual(marks)
 })
